Abstracted out ModelStub from all test classes
